added loading first page

+++ app/src/main/java/org/mixdrinks/mixdrinks/models/cocktail.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class Cocktail (
     @SerializedName("id"         ) var id         : Int?              = null,
     @SerializedName("name"       ) var name       : String?           = null,
-    @SerializedName("rating"     ) var rating     : Int?              = null,
+    @SerializedName("rating"     ) var rating     : Float?              = null,
     @SerializedName("visitCount" ) var visitCount : Int?              = null,
     @SerializedName("images"     ) var images     : ArrayList<Image> = arrayListOf()
